+++ src/components/HomePage/Project/Project.tsx
@@ -2,8 +2,7 @@ import React, { FunctionComponent, ReactElement } from 'react'
 
 import ProjectInfo from '@/components/HomePage/ProjectInfo'
 import ScrollAnimation, { fadeInDirection } from '@/components/HomePage/ScrollAnimation/ScrollAnimation'
-
-import { IProject } from '@/common/interfaces'
+import { IProject } from '@/interfaces'
 
 import './styles.css'
