+++ src/components/HomePage/Testimonials/__tests__/Testimonials.test.tsx
@@ -1,9 +1,8 @@
 import { render, screen } from '@testing-library/react'
 import React from 'react'
 
+import { ITestimonial } from '@/interfaces'
 import '@testing-library/jest-dom'
-
-import { ITestimonial } from '@/common/interfaces'
 
 import testimonials from '@/data/testimonials'
