+++ src/interfaces/IAppContext.ts
@@ -2,11 +2,11 @@ import IContact from './IContact'
 import IProject from './IProject'
 import ITestimonial from './ITestimonial'
 
-interface IContext {
+interface IAppContext {
   appVersion: string
   contacts: IContact
   projects: IProject[]
   testimonials: ITestimonial[]
 }
 
-export default IContext
+export default IAppContext
