+++ src/interfaces/IDomain.ts
