+++ src/interfaces/IMember.ts
