+++ src/interfaces/ISocialMedia.ts
