+++ src/interfaces/index.ts
@@ -1,4 +1,4 @@
-export { default as IContext } from './IContext'
+export { default as IAppContext } from './IAppContext'
 export { default as IContact } from './IContact'
 export { default as IDomain } from './IDomain'
 export { default as IMember } from './IMember'
